+++ micropython/wprest.py
@@ -1,4 +1,4 @@
-# Firmware Version: v2.03.0
+# Firmware Version: v2.04.0
 # wprest.py
 # Handles all WordPress REST API communication for TMON MicroPython device
 import settings